fix extension file
