--- src/menu.js
@@ -1,16 +1,0 @@
-import React from "react";
-
-const Menu = () => {
-    return <nav>
-        <ul>
-            <li>Premium</li>
-            <li>Support</li>
-            <li>Download</li>
-            <li> | </li>
-            <li>Sign up</li>
-            <li>Log in</li>
-        </ul>
-    </nav>
-};
-
-export default Menu;
