+++ src/menu.jsx
@@ -1,0 +1,16 @@
+const Menu = () => {
+  return (
+    <nav>
+      <ul>
+        <li>Premium</li>
+        <li>Support</li>
+        <li>Download</li>
+        <li> | </li>
+        <li>Sign up</li>
+        <li>Log in</li>
+      </ul>
+    </nav>
+  );
+};
+
+export default Menu;
